+++ src/Node.ts
@@ -2444,31 +2444,27 @@ export abstract class Node<Config extends NodeConfig = NodeConfig> {
    * @param {Number} margin.y
    * @returns {Boolean}
    * @name Konva.Node#isClientRectOnScreen
+   * @example
+   * // get index
+   * // default calculations
+   * var isOnScreen = node.isClientRectOnScreen()
+   * // increase object size (or screen size) for cases when objects close to the screen still need to be marked as "visible"
+   * var isOnScreen = node.isClientRectOnScreen({ x: stage.width(), y: stage.height() })
    */
-  isClientRectOnScreen(margin?: number | {x: number; y: number;}): boolean {
-    const _margin = 
-      margin === undefined
-        ? {x: 0, y: 0} 
-        : isNaN(margin as any)
-          ? margin as {x: number; y: number;} 
-          : {x: margin as number, y: margin as number}
-    ;
-    type Rect = {[k in 'x'|'y'|'width'|'height']: number};
-    const haveIntersection = (r1: Rect, r2: Rect) => !(
-      r2.x > r1.x + r1.width ||
-      r2.x + r2.width < r1.x ||
-      r2.y > r1.y + r1.height ||
-      r2.y + r2.height < r1.y
-    );
+  isClientRectOnScreen(
+    margin: { x: number; y: number } = { x: 0, y: 0 }
+  ): boolean {
     const stage = this.getStage();
-    if(!stage) return false;
+    if (!stage) {
+      return false;
+    }
     const screenRect = {
-      x: (-stage.x() - _margin.x) / stage.scaleX(),
-      y: (-stage.y() - _margin.y) / stage.scaleY(),
-      width: (stage.width() +  _margin.x) / stage.scaleX(),
-      height: (stage.height() +  + _margin.y) / stage.scaleY()
+      x: -margin.x,
+      y: -margin.y,
+      width: stage.width() + margin.x,
+      height: stage.height() + margin.y,
     };
-    return haveIntersection(screenRect, this.getClientRect({relativeTo: stage as any}));
+    return Util.haveIntersection(screenRect, this.getClientRect());
   }
 
   preventDefault: GetSet<boolean, this>;